AI_Summarize/requirements.txt
fastapi
uvicorn
python-dotenv
pypdf
google-generativeai
langchain
langchain-community
langchain-google-genai
python-multipart
sentence-transformers
torch
pgvector
psycopg2-binary
